dc.c robot.c farin ad nalgast

+++ Mek_Hand_Robot-main/Mek_HandGesture_Robot.X/robot.h
@@ -15,16 +15,16 @@
 #define SERVO1MIN 100
 #define SERVO1MAX 500
 
-uint8_t servo0_stat = 300;
-uint8_t servo1_stat = 300;
-uint8_t servo2_stat = 300;
-uint8_t servo3_stat = 300;
+uint16_t servo0_stat = 300;
+uint16_t servo1_stat = 300;
+uint16_t servo2_stat = 300;
+uint16_t servo3_stat = 300;
 
 
 void drive(uint8_t data_x, uint8_t data_y);
 void shoulder(uint8_t data_x, uint8_t data_y);
 void front_arm(uint8_t data_x, uint8_t data_y);
-void move_servo(uint8_t servo, uint8_t data, uint8_t *servo_stat);
+void move_servo(uint8_t servo, uint8_t data, uint16_t *servo_stat);
 void move_servo1_up(uint8_t n);
 void move_servo1_down(uint8_t n);
